Clean up previous FilterIndexRule and JoinIndexRule (#448)
